Update HTML, SCSS for home; finish '_cover.scss'
Implemented landing section buttons in the form of <a>. Added pointer
to header with '#landing:after'. Added spacing between '#landing' and
'#home-content'. Created mixin to lighten 'a.btn's on hover. Discovered
some issues with scaling down the component; the position mixins added
in the last commit did not center the second <div> in the header that
was meant to hold the text "Still confused?"