Rearrange nav, simplify header, featurettes in Construction
Nav:
1. Put '.navbar-brand' before '.navbar-toggler'
2. Add 'onclick' event to <span> in '.navbar-brand' to access homepage
3. Change links in 'ul.nav.navbar-nav' to access respective pages

Header:
The complex <header> was discarded in favor of a simpler <header> that
is easier to design and implement. Contains <h1> title and <p>
introductory text.

Featurrettes:
With current time constraints, it was necessary to simplify how the
content was displayed. Instead of three rows, the new layout contains
two rows - with two columns per each row. Each column will describe a
technology used to construct this site, with a link to each technology's
license.