Switch to single line comments in '_custom.scss'
After learning that single line comments are ignored in SASS upon
compilation (whereas multiline comments are not), each comment in the
'_custom.scss' file was switched over to a single line comment. In doing
so, the resulting 'app.css' document now has fewer unnecessary comments.
This is because '_custom.scss' is made up entirely of variables, which
are also not relevant upon compilation and therefore not added.

In the future, whether or not the comment will be useful when SASS is
compiled should be kept in mind to avoid bloated file sizes and
confusing documentation.