Simplify, extend classes in '_cover.scss'
Since a <header> class was added to '_base.scss', the <a>s were turned
to <button>s in 'index.html', and the color classes were added to
buttons ('.btn', <button>) in '_buttons.scss', a lot could be stripped
from '_cover.scss'.

What remains are simplified and extended classes that now incorporate
some media breakpoints to help transition smoothly to smaller devices.

For example, all the elements are now inside a universal
'header#landing', as this is the only element within which they are
applicable.

On smaller browser widths, the 'div.container' spans 100%, and the
buttons have an increasingly smaller amount of padding.

In addition, the 'header#landing' class now has a min-height of 350px,
to transition better if the browser height becomes too small.