Add '_jumbotron.scss' component
For all pages except home, a <header> with a '.jumbotron' is present.
Instead of the current Bootstrap SASS, a customized component was
created instead, as much less code was needed than would be provided.

In '_jumbotron.scss', a 'header.jumbotron' is added with a large
amount of top and bottom padding. Enclosed within is a 'div.container',
which has a max-width of 1000px and is centered with 'margin: 0 auto'.