Add real content, links, sources to Construction
Replaced filler material in 'construction/index.html' with actual
content, links, and sources. Each technology has a sources button that,
on click, displays in a popover what sources were used. Functionality
for the popover that was implemented in the last commit
(246cb99) was referenced as <script> dependency for 'app.js'
near the bottom of the page.