Add real content, links, sources, citations to Hosting
Replaced filler material in 'construction/index.html' with actual
content, links, and sources. Each technology has a sources button that,
on click, displays in a popover what sources were used.

In addition, each video has a citation underneath in the hopes of
preventing people from thinking that the video was created by
myself.