Change 'button.btn'; add 'div.col-lg-6' in 'featurette.scss'
Since there will now be two buttons per each technology in both
'hosting/index.html' and 'construction/index.html', buttons no longer
have to be set to 100%. On smaller screens, the first button in this
same situation will have a bottom margin of 1 rem.

The 'footer.blockquote-footer' in each 'div.col-lg-6' for
'hosting/index.html' will have a font-size of 100% when viewed on
screens larger than the 'sm' media breakpoint. This is done for
readability reasons.